i18n: Prepare for external maintenance of po/*.po
When the po/*.po are maintained outside of xnec2c (e.g. by
the Translation Project), it makes no sense to update the
*.po files all the time when running "make dist" or
"make distcheck", and then having to "git restore '*.po'"
them each and every time.

This introduces the "make update-po" target in the top-level
Makefile.am to avoid having to type "make -C po update-po".

It also makes it possible to update other things in
connection with the po/ update-po.